// src/enums/discord-timestamps.ts
export enum DiscordTimestamps {
    DEFAULT = '', // 28 November 2018 09:01
    SHORT_TIME = 't', // 9:01
    LONG_TIME = 'T', // 09:01:00
    SHORT_DATE = 'd', // 28/11/2018
    LONG_DATE = 'D', // 28 November 2018
    SHORT_DATE_TIME = 'f', // 28 November 2018 09:01
    LONG_DATE_TIME = 'F', // Wednesday, 28 November 2018 09:01
    RELATIVE = 'R', // 3 years ago
}